+++ event-sources/event-source-kafka/src/test/java/io/telicent/smart/cache/sources/kafka/TestKafkaEventSource.java
@@ -148,7 +148,7 @@ public class TestKafkaEventSource extends AbstractEventSourceTests<Integer, Stri
     @SuppressWarnings("rawtypes")
     public void givenEmptyEventList_whenDeterminingCommitOffsets_thenEmptyMap() {
         // Given
-        List<Event> events = Collections.emptyList();
+        List<Event<?,?>> events = Collections.emptyList();
 
         // When
         Map<TopicPartition, OffsetAndMetadata> offsets = KafkaEventSource.determineCommitOffsetsFromEvents(events);
@@ -157,8 +157,7 @@ public class TestKafkaEventSource extends AbstractEventSourceTests<Integer, Stri
         Assert.assertTrue(offsets.isEmpty());
     }
 
-    @SuppressWarnings("rawtypes")
-    private void generateKafkaEvents(List<Event> events, int count, String topic, int partition, long startingOffset) {
+    private void generateKafkaEvents(List<Event<?,?>> events, int count, String topic, int partition, long startingOffset) {
         for (int i = 0; i < count; i++) {
             ConsumerRecord<Long, String> record =
                     new ConsumerRecord<>(topic, partition, startingOffset + i, startingOffset + i,
@@ -168,7 +167,7 @@ public class TestKafkaEventSource extends AbstractEventSourceTests<Integer, Stri
     }
 
     @SuppressWarnings("rawtypes")
-    private void generateSimpleEvents(List<Event> events, int count) {
+    private void generateSimpleEvents(List<Event<?,?>> events, int count) {
         for (long i = 0; i < count; i++) {
             events.add(new SimpleEvent<>(Collections.emptyList(), i, "Event #" + i));
         }
@@ -178,7 +177,7 @@ public class TestKafkaEventSource extends AbstractEventSourceTests<Integer, Stri
     @SuppressWarnings("rawtypes")
     public void givenKafkaEventList_whenDeterminingCommitOffsets_thenExpectedOffsets() {
         // Given
-        List<Event> events = new ArrayList<>();
+        List<Event<?,?>> events = new ArrayList<>();
         generateKafkaEvents(events, 100, KafkaTestCluster.DEFAULT_TOPIC, 0, 0);
         generateKafkaEvents(events, 50, KafkaTestCluster.DEFAULT_TOPIC, 1, 50);
         generateKafkaEvents(events, 10, KafkaTestCluster.DEFAULT_TOPIC, 2, 100);
@@ -198,7 +197,7 @@ public class TestKafkaEventSource extends AbstractEventSourceTests<Integer, Stri
     @SuppressWarnings("rawtypes")
     public void givenNonKafkaEventList_whenDeterminingCommitOffsets_thenNoOffsets() {
         // Given
-        List<Event> events = new ArrayList<>();
+        List<Event<?,?>> events = new ArrayList<>();
         generateSimpleEvents(events, 100);
 
         // When
@@ -212,7 +211,7 @@ public class TestKafkaEventSource extends AbstractEventSourceTests<Integer, Stri
     @SuppressWarnings("rawtypes")
     public void givenMixedEventList_whenDeterminingCommitOffsets_thenExpectedOffsets() {
         // Given
-        List<Event> events = new ArrayList<>();
+        List<Event<?,?>> events = new ArrayList<>();
         generateKafkaEvents(events, 100, KafkaTestCluster.DEFAULT_TOPIC, 0, 0);
         generateKafkaEvents(events, 50, KafkaTestCluster.DEFAULT_TOPIC, 1, 200);
         generateSimpleEvents(events, 100);